+++ src/test/java/org/opentripplanner/gtfs/GtfsContextBuilder.java
@@ -43,8 +43,7 @@ public class GtfsContextBuilder {
         GtfsImport gtfsImport = gtfsImport(defaultFeedId, path);
         GtfsFeedId feedId = gtfsImport.getFeedId();
         OtpTransitServiceBuilder transitBuilder = mapGtfsDaoToInternalTransitServiceBuilder(
-                gtfsImport.getDao(),
-                new DataImportIssueStore(false)
+                gtfsImport.getDao(), feedId.getId(), new DataImportIssueStore(false)
         );
         return new GtfsContextBuilder(
                 feedId,