+++ src/test/java/org/opentripplanner/model/impl/OtpTransitServiceBuilderTest.java
@@ -103,7 +103,7 @@ public class OtpTransitServiceBuilderTest {
 
     private static FareAttribute createFareAttribute(Agency agency) {
         FareAttribute fa = new FareAttribute();
-        fa.setId(new FeedScopedId(agency.getId(), "FA"));
+        fa.setId(new FeedScopedId(FEED_ID, "FA"));
         return fa;
     }